requirements.txt
numpy >= 1.26.3
scipy >= 1.11.4
pandas >= 2.1.4
matplotlib >= 3.8.2
seaborn >= 0.13.1
hydra-core >= 1.3.2
setuptools >= 68.2.2
cma == 3.4.0
